Add Contact.java and Update Address.java, Email.java , Phone.java
